requirements.txt
requests==2.28.1
flask==2.2.2
google-cloud-firestore==2.5.3